# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BaaS Dashboard API",
    "description": "\n# Bot-as-a-Service Dashboard API\n\nAPI completa para gerenciamento de bots conversacionais multi-canal.\n\n## Autenticação\n\nTodas as rotas requerem autenticação via JWT (NextAuth) ou API Key.\n\n```\nAuthorization: Bearer <token>\n```\n\n## Rate Limiting\n\n- **Free tier**: 100 requests/minute\n- **Pro tier**: 1000 requests/minute  \n- **Enterprise**: Custom limits\n\nHeaders de resposta incluem:\n- `X-RateLimit-Limit`: Limite de requests\n- `X-RateLimit-Remaining`: Requests restantes\n- `Retry-After`: Segundos para retry (quando limitado)\n\n## Errors\n\nTodos os erros seguem o formato:\n\n```json\n{\n  \"error\": {\n    \"message\": \"Human-readable message\",\n    \"code\": \"MACHINE_CODE\",\n    \"details\": {}\n  }\n}\n```\n\n## Paginação\n\nEndpoints que retornam listas suportam paginação:\n\n- `page`: Número da página (1-indexed)\n- `limit`: Itens por página (max 100)\n\nResposta inclui objeto `pagination`:\n\n```json\n{\n  \"pagination\": {\n    \"page\": 1,\n    \"limit\": 20,\n    \"total\": 45,\n    \"totalPages\": 3\n  }\n}\n```\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "BaaS Support",
      "email": "support@baas.example.com",
      "url": "https://baas.example.com/docs"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
...
    },
    {
...
  "paths": {
...
    "/api/clawdbot/webhook": {
      "get": {
        "tags": [
          "Clawdbot"
        ],
        "summary": "Webhook health check",
        "description": "Returns webhook service status",
        "operationId": "getClawdbotWebhookStatus",
        "responses": {
          "200": {
            "description": "Service status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "service": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clawdbot"
        ],
        "summary": "Receive webhook event",
        "description": "Receives events from Clawdbot Gateway (messages, status changes, etc.)",
        "operationId": "receiveClawdbotWebhook",
        "parameters": [
          {
            "name": "x-clawdbot-signature",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "HMAC-SHA256 signature for payload verification"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "received": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "code": {
                "type": "string",
                "description": "Machine-readable error code"
              },
              "details": {
                "oneOf": [
                  {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  {
                    "type": "object"
                  },
                  {
                    "type": "string"
                  }
                ],
                "description": "Additional error details"
              }
            }
          }
        },
        "example": {
          "error": {
            "message": "Validation error",
            "code": "VALIDATION_ERROR",
            "details": [
              {
                "path": "name",
                "message": "Required"
              }
            ]
          }
        }
      },
...
      "WebhookEvent": {
        "type": "object",
        "required": [
          "type",
          "timestamp"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "message.received",
              "message.sent",
              "group.joined",
              "group.left",
              "status.change",
              "agent.response"
            ]
          },
          "timestamp": {
            "type": "integer",
            "description": "Unix timestamp in milliseconds"
          }
        },
        "discriminator": {
          "propertyName": "type"
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from NextAuth session"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API key for external integrations"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "message": "Unauthorized",
                "code": "UNAUTHORIZED"
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "message": "Internal server error",
                "code": "INTERNAL_ERROR"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}